refactor: Move ModelWrapper under models
